requirements.txt
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
fastapi>=0.103.0
uvicorn>=0.23.2
pyyaml>=6.0.1
